+++ app/actions/get-next-events.tsx
@@ -2,11 +2,11 @@
 
 import { getEventsPaginated } from '@/lib/get-events-paginated';
 
-export async function getNextEvents(page: number, nbPerPage: number, organizerId?: string) {
+export async function getNextEvents(page: number, nbPerPage: number, organizerId?: string, title?: string) {
 	let eventsPaginated;
 
 	try {
-		eventsPaginated = await getEventsPaginated(page, nbPerPage, organizerId);
+		eventsPaginated = await getEventsPaginated(page, nbPerPage, organizerId, title);
 	} catch (error) {
 		return null;
 	}
